+++ algorithm/superset_zeta_transform.cpp
@@ -1,14 +1,13 @@
 #pragma once
-#include <vector>
 #include <cassert>
-#include "../utility/int_alias.cpp"
+#include <vector>
 #include "../utility/rep.cpp"
 
 template <class T> void superset_zeta_transform(std::vector<T>& f) {
-    const usize n = f.size();
+    const int n = f.size();
     assert((n & (n - 1)) == 0);
-    for (usize i = 1; i < n; i <<= 1) {
-        for (const usize j : rep(0, n)) {
+    for (int i = 1; i < n; i <<= 1) {
+        for (const int j : rep(0, n)) {
             if (j & i) f[j & ~i] += f[j];
         }
     }
